Avoid deadlock in event start
When using streaming, the main event loop is blocked to await tokens, if we try to
execute a function (Message.send() here) in the same event loop, it leads to a deadlock.
Adding a new_loop parameter to the run_sync function allows to create a new event loop
in the current Thread and allows to send a Message during a stream.