Gracefully handle HTTP errors when sending steps
Fixes #1232

Modify Chainlit so that HTTP errors which occur while sending steps
to LiteralAI are caught and logged.

Prior to this change, HTTP errors such as timeouts result in a series
of cascading exceptions and tracebacks that begin with "Task exception
was never retrieved" and continue for several traceback iterations.
The result is a very verbose log which is hard to understand.

With this change we will see one-line error messages in the log
rather than a series of tracebacks.

In the future we might want to improve this further by retrying
(sending the steps again) when the HTTP error is temporary.