fix: Prevent constant refresh on the Login page (#2205)
When remaining on the Login page, users can notice the page constantly
refreshing (approximately every 5 seconds). This is caused by SWR
(stale-while-revalidate) which automatically retries the /user route on
error (in this case, a 401), which causes the user to be set to
undefined in userManagement.ts while fetching the latest status. Having
an undefined user results in a full page loading indicator, making it
look like Chainlit is reloading.

The main part of the fix is to not set the user to undefined while
waiting for the /user route.

In addition to this, I've added an await for the handleAuth call in the
Login page so that we properly keep the controls disabled while waiting
for the authentication result (e.g. time taken by the auth_callback).

Finally, to further improve the user experience, I've disabled the
username and password fields in addition to the login button while
logging in.

Here are the updates in action:


![chainlit_login_improvements](https://github.com/user-attachments/assets/369c205d-41fb-48c2-ad43-ea04f0734f03)

---------

Co-authored-by: Josh Hayes <35790761+hayescode@users.noreply.github.com>
Co-authored-by: Aleksandr Vishniakov <asvishnyakov@hotmail.com>