Add a second optional language parameter to set_chat_profiles (#2461)
_As suggested by @sandangel in
https://github.com/Chainlit/chainlit/pull/2456._

This changes adds an optional language parameter for set_chat_profiles
callback, in addition to the currently optional User parameter.

This allows for programmatically localizing the label and description
associated to the chat profiles.

Includes a test to confirm the feature works as expected.

⚠️ I am not sure about the changes in session.py where we get the
language from the HTTP_ACCEPT_LANGUAGE header, and I fear we might have
cases where the language will not match the one passed in the
/project/settings route.

---------

Co-authored-by: Josh Hayes <35790761+hayescode@users.noreply.github.com>
Co-authored-by: Aleksandr Vishniakov <asvishnyakov@hotmail.com>
Co-authored-by: San Nguyen <22189661+sandangel@users.noreply.github.com>