refactor: allow Thread Sharing by user-defined callback (#2521)
I want to allow admins to view any threads. I determine if a user is an
admin using custom RBAC roles and saving in the user.metadata.
Presumably other developers will have other ways of doing this.

This change makes it so that as long as the `@cl.on_shared_thread_view`
returns True then the thread can be viewed, instead of requiring the
user to share the thread. Developers can customize the logic within this
callback however they see fit.