fix: on_chat_start not always firing (#2549)
fixes #2535

This bug only happens **some of the time** and causes code inside
on_chat_start, like chat_settings and LLM configuration, to not fire.
Clicking 'New Chat' is the only sure-fire way to trigger on_chat_start.

After lots of debugging trying to reproduce this bug I noticed that
`context.session.restored` conditional was hit which skips
on_chat_start. I was able to fix this issue by adding an additional
conditional `context.session.has_first_interaction` which can only be
True when the user has actually interacted.

Like mentioned it's difficult to reliably reproduce this bug, and the
root cause is still alluding me, but for the meantime this does appear
to fix the issue.