ci: libs publishing workflow (#2570)
Closes #2550

This PR enables automatic publishing of the `@chainlit/react-client`
(and any other non-private package in `libs/*` if we will add them) to
the **npm** registry.

To test the publishing process, use the **Dry run (test publishing)**
option, similar to the **Publish to TestPyPI instead of PyPI** feature
introduced in #2483.

### Other changes:

- Removed `pnpm install` arguments, since CI **must always use a frozen
lockfile**.
* **Updating the lockfile during CI runs is bad practice**, as changes
are not committed back to the repository and resolved dependency
versions may differ from those declared in the source code.
* If you encounter an error where `pnpm` indicates that the lockfile
needs to be updated, run `pnpm install` or `pnpm install
--lockfile-only` locally and commit the resulting changes to the PR
where the CI failure occurred.
- Removed `pnpm-skip-install` from `pnpm-node-install` action since no
longer used anywhere
- CI will be skipped if the PR or commit includes only markdown files
located at the repository root or the `LICENSE` file.