Bug: Fix page blinking after login (#2640)
When using the `header_auth_callback ` and another auth callback for
login, the page starts to "blink" as soon as a user logged in. This is
because `header_auth_callback ` always gets called when routing to the
root.

Fix: Now when a user is already authenticated, the `header_auth_callback
`won't be called
Issue: #2599